+++ 3d.rkt
@@ -613,6 +613,7 @@ cutaway3d
  trajectory-ribbon3d
  trajectory-bundle3d
  trajectory-inspection3d
+ trajectory-pick-inspection3d
  equilibrium-inspection3d
  linearization-inspection3d
  flow-map-inspection3d